implementing screening time overlaps v1

+++ src/main/java/by/tami/kinotower/hall/repository/HallRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface HallRepository extends JpaRepository<Hall, Integer> {
+public interface HallRepository extends JpaRepository<Hall, Long> {
     boolean existsByName(String name);
 
     @Query("""
